ecommerce-backend/ai-services/requirements.txt
# Core ML Libraries
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
scipy==1.11.1

# Deep Learning
tensorflow==2.13.0
torch==2.0.1

# NLP
transformers==4.30.2
sentence-transformers==2.2.2
spacy==3.6.0

# API Framework
flask==2.3.2
flask-cors==4.0.0

# Database
psycopg2-binary==2.9.6
redis==4.6.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
joblib==1.3.1

# OpenAI (for GPT features)
openai==0.27.8
anthropic==0.3.0